refactor: Refactor structure of heroes folder
