+++ src/app/heroes/hero-detail/hero-detail.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { Location } from '@angular/common';
 
-import { Hero } from '../hero';
+import { Hero } from '../../core/interface/hero';
 import { HeroService } from '../hero.service';
 
 @Component({
