+++ src/app/heroes/hero.service.ts
@@ -1,10 +1,8 @@
 import { Injectable } from '@angular/core';
-
 import { Observable, of } from 'rxjs';
-
-import { Hero } from './hero';
 import { HEROES } from './mock-heroes';
-import { MessageService } from './message.service';
+import { MessageService } from '../message.service';
+import { Hero } from '../core/interface/hero';
 
 @Injectable({ providedIn: 'root' })
 export class HeroService {
