+++ src/app/heroes/heroes.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 
-import { Hero } from '../hero';
-import { HeroService } from '../hero.service';
+import { Hero } from '../core/interface/hero';
+import { HeroService } from './hero.service';
 
 @Component({
   selector: 'app-heroes',
